rearrange content
